[knx] Fix write blocked forever after read from bus
Carryover from smarthomej/addons#299 and smarthomej/addons#330.

Also-by: Jan N. Klug <github@klug.nrw>
